@@ -18,16 +18,16 @@ publish_to: 'none'
 version: 1.0.2+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  percent_indicator: ^2.1.9
-  flutter_hooks: ^0.15.0
-  responsive_builder: ^0.3.0
+  percent_indicator: ^3.3.0-nullsafety.1
+  flutter_hooks: ^0.16.0
+  responsive_builder: ^0.4.0-nullsafety.1
 
 dev_dependencies:
   flutter_test:
