@@ -1,7 +1,7 @@
 class Consts {
-  static const neutralState = 0;
-  static const correctState = 1;
-  static const halfCorrectState = -1;
-  static const incorrectState = -2;
+  static const String neutralState = "0";
+  static const String correctState = "1";
+  static const String halfCorrectState = "-1";
+  static const String incorrectState = "-2";
   static const String returnSymbol = "\u23CE";
 }
